[Core] Move TimeService.Stub to the test source.
Extract TimeService.Stub to TimeServiceStub and move it to the test
source.
