+++ core/src/main/java/cucumber/runner/TimeService.java
@@ -10,20 +10,4 @@ public interface TimeService {
         }
     };
 
-    class Stub implements TimeService {
-        private final long duration;
-        private final ThreadLocal<Long> currentTime = new ThreadLocal<Long>();
-
-        public Stub(long duration) {
-            this.duration = duration;
-        }
-
-        @Override
-        public long time() {
-            Long result = currentTime.get();
-            result = result != null ? result : 0l;
-            currentTime.set(result + duration);
-            return result;
-        }
-    }
 }
